requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mysql-connector-python==8.2.0
bcrypt==4.1.2
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
protobuf==3.20.3
mediapipe==0.10.7
